functions/discover_pairs/requirements.txt
google-cloud-bigquery-storage>=2.24.0
google-cloud-storage>=2.5.0
google-cloud-bigquery>=3.14.0
google-cloud-pubsub>=2.17.0
pandas>=2.2.0
db-dtypes>=1.2.0  
